+++ evalml/automl/auto_base.py
@@ -259,7 +259,8 @@ class AutoBase:
             objectives_to_score = [self.objective] + self.additional_objectives
             try:
                 pipeline.fit(X_train, y_train, self.objective)
-                score, other_scores = pipeline.score(X_test, y_test, objectives=objectives_to_score)
+                scores = pipeline.score(X_test, y_test, objectives=objectives_to_score)
+                score = scores[self.objective.name]
                 plot_data.append(pipeline.get_plot_data(X_test, y_test, self.plot_metrics))
             except Exception as e:
                 if raise_errors:
@@ -267,10 +268,10 @@ class AutoBase:
                 if pbar:
                     pbar.write(str(e))
                 score = np.nan
-                other_scores = OrderedDict(zip([n.name for n in self.additional_objectives], [np.nan] * len(self.additional_objectives)))
+                scores = OrderedDict(zip([n.name for n in self.additional_objectives], [np.nan] * len(self.additional_objectives)))
             ordered_scores = OrderedDict()
             ordered_scores.update({self.objective.name: score})
-            ordered_scores.update(other_scores)
+            ordered_scores.update(scores)
             ordered_scores.update({"# Training": len(y_train)})
             ordered_scores.update({"# Testing": len(y_test)})
             cv_data.append({"all_objective_scores": ordered_scores, "score": score})
@@ -281,7 +282,8 @@ class AutoBase:
         self._add_result(trained_pipeline=pipeline,
                          parameters=parameters,
                          training_time=training_time,
-                         cv_data=cv_data)
+                         cv_data=cv_data,
+                         plot_data=plot_data)
 
         desc = "✔" + desc[1:]
         pbar.set_description_str(desc=desc, refresh=True)
@@ -297,7 +299,7 @@ class AutoBase:
         proposal = zip(space, values)
         return list(proposal)
 
-    def _add_result(self, trained_pipeline, parameters, training_time, cv_data):
+    def _add_result(self, trained_pipeline, parameters, training_time, cv_data, plot_data):
         scores = pd.Series([fold["score"] for fold in cv_data])
         score = scores.mean()
 
@@ -323,7 +325,8 @@ class AutoBase:
             "score": score,
             "high_variance_cv": high_variance_cv,
             "training_time": training_time,
-            "cv_data": cv_data
+            "cv_data": cv_data,
+            "plot_data": plot_data
         }
 
         self.results['search_order'].append(pipeline_id)
@@ -381,10 +384,6 @@ class AutoBase:
 
         all_objective_scores = [fold["all_objective_scores"] for fold in pipeline_results["cv_data"]]
         all_objective_scores = pd.DataFrame(all_objective_scores)
-
-        # note: we need to think about how to better handle metrics we don't want to display in our chart
-        # currently, just dropping the columns before displaying
-        all_objective_scores = all_objective_scores.drop(["ROC", "Confusion Matrix"], axis=1, errors="ignore")
 
         for c in all_objective_scores:
             if c in ["# Training", "# Testing"]: