+++ evalml/objectives/standard_metrics.py
@@ -1,3 +1,5 @@
+from abc import ABC
+
 import numpy as np
 import pandas as pd
 from sklearn import metrics
@@ -316,8 +318,12 @@ class ExpVariance(RegressionObjective):
         return metrics.explained_variance_score(y_true, y_predicted)
 
 
-class PlotMetric:
+class PlotMetric(ABC):
     score_needs_proba = True
+    name = None
+
+    def score(self, y_predicted, y_true):
+        raise NotImplementedError("score() is not implemented!")
 
 
 class ROC(PlotMetric):
@@ -332,8 +338,6 @@ class ROC(PlotMetric):
 
 class ConfusionMatrix(PlotMetric):
     """Confusion matrix for classification problems"""
-    needs_fitting = False
-    greater_is_better = True
     score_needs_proba = False
     name = "Confusion Matrix"
     problem_types = [ProblemTypes.BINARY, ProblemTypes.MULTICLASS]