+++ evalml/tests/pipeline_tests/regression_pipeline_tests/test_linear_regression.py
@@ -39,13 +39,18 @@ def test_linear_regression(X_y_categorical_regression):
                                    fit_intercept=True,
                                    random_state=0,
                                    n_jobs=-1)
-    clf.fit(X, y, objective)
-    clf_score = clf.score(X, y, [objective])
+    clf.fit(X, y)
+    clf_scores = clf.score(X, y, [objective])
     y_pred = clf.predict(X)
 
     np.testing.assert_almost_equal(y_pred, sk_pipeline.predict(X), decimal=5)
-    np.testing.assert_almost_equal(sk_score, clf_score[0], decimal=5)
+    np.testing.assert_almost_equal(sk_score, clf_scores[objective.name], decimal=5)
     assert not clf.feature_importances.isnull().all().all()
+
+    # testing objective parameter passed in does not change results
+    clf.fit(X, y, objective)
+    y_pred_with_objective = clf.predict(X, objective)
+    assert((y_pred == y_pred_with_objective).all())
 
 
 def test_lr_input_feature_names(X_y):